ok, the merge really flubbed. clean up and restart.
this is a result of manually copying files over from the 0.17.7
tree into the directory, instead of merging it, while keeping our
changes around.

it compiles now under x86 and riscv32, but the rust2c generation code
seems broken due to a missing include path, but let's commit
this here before digging into that.